requirements.txt
wave
yt-dlp
python-dotenv
google-api-python-client
requests